Python/requirements.txt
aiohttp==1.3.1
async-timeout==1.1.0
chardet==2.3.0
multidict==2.1.4
python-engineio==3.8.1
python_socketio==4.1.0
six==1.10.0
yarl==0.9.2
